Datatype conversion update

+++ QueryDB.Core.Tests/Entities.cs
@@ -149,7 +149,7 @@ namespace QueryDB.Core.Tests
                 public DateTime Date_Column { get; set; }
                 public decimal Float_Column { get; set; }
                 public decimal Integer_Column { get; set; }
-                public int IntervalYearToMonth_Column { get; set; }
+                public long IntervalYearToMonth_Column { get; set; }
                 public TimeSpan InternalDayToSecond_Column { get; set; }
                 public string Long_Column { get; set; }
                 public byte[] NChar_Column { get; set; }
